requirements.txt
apache-airflow-providers-amazon
pyspark
boto3